Updated docs

+++ spring-cloud-contract-stub-runner/src/main/java/org/springframework/cloud/contract/stubrunner/ClasspathStubProvider.java
@@ -69,6 +69,9 @@ public class ClasspathStubProvider implements StubDownloaderBuilder {
 				if (log.isDebugEnabled()) {
 					log.debug("For paths " + paths + " found following resources " + resources);
 				}
+				if (resources.isEmpty()) {
+					throw new IllegalStateException("No stubs were found on classpath for [" + config.getGroupId() + ":" + config.getArtifactId() + "]");
+				}
 				final File tmp = createTempDir();
 				tmp.deleteOnExit();
 				Pattern groupAndArtifactPattern = Pattern.compile(
